requirements.txt
flask
flask-classful
flask-security
flask-admin
gunicorn
sentry-sdk[flask]
py-healthcheck
sqlalchemy
flask-sqlalchemy
flask-cors
requests
psycopg2-binary
email_validator
minio
flask-minio
bleach